pyproject.toml
[tool.poetry]
name = "tracenexus"
version = "0.1.7"
description = "A platform-agnostic MCP server for LLM observability"
authors = ["Rajesh Sharma <sharrajesh@gmail.com>"]
readme = "README.md"
packages = [{include = "tracenexus"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastmcp = "^2.3.4"
langchain-mcp-adapters = "^0.1.0"
pydantic = "^2.11.4"
typing-extensions = "^4.13.2"
langsmith = "^0.3.42"
langfuse = "^2.60.5"
aiohttp = "^3.11.18"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
mypy = "^1.15.0"
isort = "^6.0.1"
ruff = "^0.11.10"
pytest-cov = "*"
types-pyyaml = "^6.0.12.20250516"
keyring = "^25.6.0"
keyrings-alt = "^5.0.2"

[tool.poetry.scripts]
tracenexus = "tracenexus.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "langfuse.*"
ignore_missing_imports = true
